services/inference-service/requirements.txt
ultralytics
grpcio
grpcio-tools
protobuf
opencv-python-headless
numpy
confluent-kafka
pydantic
